🚚 Move JSON schema related files
