+++ src/index.ts
@@ -1,5 +1,2 @@
 export * as shopify from './shopify'
-export { createTemplateSchema, TemplateJsonSchema } from './template'
-export { createSectionSchema, SectionJsonSchema } from './section'
-export { createBlockSchema, BlockJsonSchema } from './block'
-export { createSettingSchema, SettingJsonSchema } from './setting'
+export * as jsonSchema from './transformer'
